requirements.txt
flask
pandas
gunicorn
sklearn
scikit-image
tensorflow
tensorflow.keras.preprocessing
skimage.io
numpy
